email retabli en dev

+++ src/components/Carousel/utils.ts
@@ -8,7 +8,7 @@
  */
 export function formatDateWithDay(dateStr: string): string {
   const options: Intl.DateTimeFormatOptions = {
-    weekday: "long", // "jeu."
+    weekday: "short", // "jeu."
     year: "numeric",  // "2025"
     month: "long",    // "janvier"
     day: "numeric",   // "16"
@@ -34,7 +34,7 @@ export function formatDateRange(startStr: string, endStr: string): string {
     // -> "jeu. 16" (sans mois/année) + "et" + "ven. 17 janvier 2025"
     // 1) Format "jour + nom du jour" (sans mois/année)
     const shortOptions: Intl.DateTimeFormatOptions = {
-      weekday: "long",
+      weekday: "short",
       day: "numeric",
     };
     const startShort = startDate.toLocaleDateString("fr-FR", shortOptions);
